Dev's projects are now auto-rescheduled when a new one override the existing ones

+++ db/migrate/20110311194653_create_projects.rb
@@ -7,7 +7,7 @@ class CreateProjects < ActiveRecord::Migration
       t.date :started_on
       t.date :ended_on
       t.integer :estimated_duration
-      t.float :actual_duration
+      t.integer :actual_duration
 
       t.timestamps
     end
